modifico int por float

+++ script_creacion_inicial.sql
@@ -157,7 +157,7 @@ CREATE TABLE [GAME_OF_CODE].[Sucursal] (
 CREATE TABLE [GAME_OF_CODE].[Detalle_Factura] (
 	[id_detalle_factura] INT IDENTITY(1,1) PRIMARY KEY,
 	[item_factura] [nvarchar](50) NOT NULL DEFAULT 'Producto',
-	[item_monto] INT NOT NULL,
+	[item_monto] FLOAT NOT NULL,
 	[cantidad] INT NOT NULL,
 	[id_factura] INT NOT NULL
 )
@@ -165,7 +165,7 @@ CREATE TABLE [GAME_OF_CODE].[Detalle_Factura] (
 CREATE TABLE [GAME_OF_CODE].[Pago_de_Facturas] (
 	[id_pago_facturas] INT IDENTITY(1,1) PRIMARY KEY,
 	[fecha_cobro] [datetime] NOT NULL,
-	[importe] INT NOT NULL,
+	[importe] FLOAT NOT NULL,
 	[id_sucursal] INT NOT NULL,
 	[id_medio_pago] INT
 )
@@ -174,7 +174,7 @@ CREATE TABLE [GAME_OF_CODE].[Factura] (
 	[id_factura] INT IDENTITY(1,1) PRIMARY KEY,
 	[numero_factura] INT NOT NULL,
 	[fecha_alta] [datetime] NOT NULL,
-	[monto_total] INT NOT NULL,
+	[monto_total] FLOAT NOT NULL,
 	[fecha_vencimiento] [datetime] NOT NULL,
 	[id_cliente] INT NOT NULL,
 	[id_empresa] INT NOT NULL,
@@ -205,8 +205,8 @@ CREATE TABLE [GAME_OF_CODE].[Cliente] (
 CREATE TABLE [GAME_OF_CODE].[Rendicion] (
 	[id_rendicion] INT IDENTITY(1,1) PRIMARY KEY,
 	[fecha_rendicion] [datetime] NOT NULL,
-	[total_rendicion] INT NOT NULL,
-	[porcentaje_comision] INT NOT NULL,
+	[total_rendicion] FLOAT NOT NULL,
+	[porcentaje_comision] FLOAT NOT NULL,
 	[importe_comision] FLOAT NOT NULL,
 	[cant_facturas_rendidas] INT NOT NULL
 )
@@ -227,7 +227,7 @@ CREATE TABLE [GAME_OF_CODE].[Empresa] (
     [nombre] [nvarchar](255) NOT NULL,
     [emp_cuit] [nvarchar](50) UNIQUE NOT NULL,
     [emp_direccion] [nvarchar](255) NOT NULL,
-	[porcentaje_comision] INT NOT NULL,
+	[porcentaje_comision] FLOAT NOT NULL,
 	[id_rubro] INT NOT NULL,
 	[estado_habilitacion] [bit] NOT NULL DEFAULT 1
 )
@@ -456,7 +456,7 @@ CREATE PROCEDURE GAME_OF_CODE.pr_crear_empresa
 	@emp_cuit nvarchar(50),
 	@emp_direccion nvarchar(255),
 	@id_rubro int,
-	@porcentaje_comision int,
+	@porcentaje_comision float,
 	@id int OUTPUT
 AS
 BEGIN
@@ -474,7 +474,7 @@ CREATE PROCEDURE GAME_OF_CODE.pr_modificar_empresa
 	@emp_direccion nvarchar(255),
 	@id_rubro int,
 	@id int, /*id_empresa*/
-	@porcentaje_comision int,
+	@porcentaje_comision float,
 	@id_output int OUTPUT
 AS
 BEGIN
@@ -488,7 +488,7 @@ GO
 CREATE PROCEDURE GAME_OF_CODE.pr_crear_factura
 	@numero_factura int,
 	@fecha_alta DATETIME,
-	@monto_total int,
+	@monto_total float,
 	@fecha_vencimiento DATETIME,
 	@id_cliente int,
 	@id_empresa int,
@@ -506,7 +506,7 @@ GO
 CREATE PROCEDURE GAME_OF_CODE.pr_modificar_factura
 	@numero_factura int,
 	@fecha_alta DATETIME,
-	@monto_total int,
+	@monto_total float,
 	@fecha_vencimiento DATETIME,
 	@id_cliente int,
 	@id_empresa int,
@@ -523,7 +523,7 @@ GO
 
 CREATE PROCEDURE GAME_OF_CODE.pr_crear_item_factura
 	@item_factura nvarchar(50),
-	@item_monto int,
+	@item_monto float,
 	@cantidad int,
 	@id_factura int,
 	@id int output
@@ -539,7 +539,7 @@ GO
 
 CREATE PROCEDURE GAME_OF_CODE.pr_modificar_item_factura
 	@item_factura nvarchar(50),
-	@item_monto int,
+	@item_monto float,
 	@cantidad int,
 	@id_factura int,
 	@id int,
@@ -585,7 +585,7 @@ GO
 
 CREATE PROCEDURE GAME_OF_CODE.pr_crear_pago_factura
 	@fecha_cobro datetime,
-	@importe int,
+	@importe float,
 	@id_sucursal int,
 	@id_medio_pago int,
 	@id int output
@@ -648,8 +648,8 @@ GO
 
 CREATE PROCEDURE GAME_OF_CODE.pr_crear_rendicion
 	@fecha_rendicion DATETIME,
-	@total_rendicion int,
-	@porcentaje_comision int,
+	@total_rendicion float,
+	@porcentaje_comision float,
 	@importe_comision float,
 	@cant_facturas_rendidas int,
 	@id int OUTPUT
